requirements.txt
# Kira V1 - Required Packages
telethon>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
yfinance>=0.2.28
gtts>=2.3.2
